change: move ExecuteTemplate to template package
